make `StagingBuffer` creation a method
